Enable shared libperl
There are packages that will use perl embedded into their applications,
for that, these packages/applications will use a shared libperl and make
use of some API designated to that use (i.e perl_alloc(), perl_construct(),
perl_free(), etc).

This patch enables the build of libperl (with the use of -Duseshrplib
configuration argument). As a result, a libperl.so will be installed and
have the required and mentioned symbols exported.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>