Add support for pre-existing Perl modules
Finds and adds support for Perl modules installed under previous
versions of Perl, so that not everything has to be upgraded
simultaneously. Has no effect in this version because 5.28.2 == 5.28.2,
but will add the 5.28.2 vendor- and site-perl paths to @INC when we
upgrade to 5.30. This will allow both modules we provide and those
installed manually by end-users to continue to be used prior to
reinstalling them.

Also removes a copy-paste leftover parameter -Dotherlibdirs that's no
longer needed.